+++ CHANGELOG.md
@@ -37,6 +37,7 @@
 ### BUG FIXES:
 
 - fix(cloudfiles_logging): ensure that the `format` attribute can't be set to null ([#93](https://github.com/fastly/terraform-provider-fastly-beta/pull/93))
+- fix(docs): use versionless domains in examples, show the nested `domain` block as an alternative on the TLS pages, and note that classic domains are only available on accounts created before September 16, 2025 ([#121](https://github.com/fastly/terraform-provider-fastly-beta/pull/121))
 - fix(docs): correct the README list of explicit/default resources, and map each one to its automatic-family equivalent ([#122](https://github.com/fastly/terraform-provider-fastly-beta/pull/122))
 
 ### Dependencies: